(#665) Refactor caching
